refactor: handshake tidyup
